[perf] : check file exist

+++ install.sh
@@ -8,18 +8,32 @@ DOTFILE_PATH=$(pwd)
     
 # fi
 
-
 # vim 配置
-ln -s $DOTFILE_PATH/.vimrc ~/.vimrc
-mkdir -p ~/.vim/syntax
-cp $DOTFILE_PATH/c.vim ~/.vim/syntax/c.vim
+vimrc_path=~/.vimrc
 
-ln -s $DOTFILE_PATH/.vim ~/.vim
-vim -c 'source ~/.vimrc | PlugInstall' -c ':qall'
+if [ ! -e "$vimrc_path" ]; then
+    echo "Creating $vimrc_path..."
+    ln -s $DOTFILE_PATH/.vimrc $vimrc_path
+    mkdir -p ~/.vim/syntax
+    cp $DOTFILE_PATH/c.vim ~/.vim/syntax/c.vim
+    ln -s $DOTFILE_PATH/.vim ~/.vim
+    vim -c 'source ~/.vimrc | PlugInstall' -c ':qall'
+else
+    echo "$vimrc_path already exists."
+fi
+
 
 # tmux 配置
-ln -s $DOTFILE_PATH/.tmux.conf ~/.tmux.conf
-tmux source-file ~/.tmux.conf
+tmux_conf_path=~/.tmux.conf
+
+if [ ! -e "$tmux_conf_path" ]; then
+    echo "Creating $tmux_conf_path..."
+    ln -s $DOTFILE_PATH/.tmux.conf $tmux_conf_path
+    tmux source-file $tmux_conf_path
+else
+    echo "$tmux_conf_path already exists."
+fi
+
 
 # bashrc
 #!/bin/bash
@@ -28,8 +42,15 @@ source ~/.bashrc
 
 
 # gdbinit
-mkdir -p ~/.config/gdb
-ln -s $DOTFILE_PATH/gdbinit ~/.config/gdb/gdbinit/
+gdb_init_path=~/.config/gdb/gdbinit/
+
+if [ ! -e "$gdb_init_path" ]; then
+    echo "Creating $gdb_init_path..."
+    mkdir -p ~/.config/gdb
+    ln -s $DOTFILE_PATH/gdbinit $gdb_init_path
+else
+    echo "$gdb_init_path already exists."
+fi
 
 
 # docker
